+++ docs/main.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { App } from './App'
+import '../src/styles/themes.css'
 import './docs.css'
 
 createRoot(document.getElementById('root')!).render(